updated with logger-module
