+++ excel-analysis.py
@@ -6,11 +6,10 @@ import logging as logger
 logger.basicConfig(filename='app.log', level=logger.INFO, format='%(asctime)s - %(message)s')
 
 
-try:
-    # Define variables for file paths
-    input_file_path = r'D:\heart.csv'
-    output_file_path = r'D:\heart-excel.xlsx'
-
+def read_csv_and_create_excel(input_file_path, output_file_path):
+    """
+    Read the first 50 rows of the CSV file, write to an Excel file, and log information.
+    """
     # Debug: Output the paths for verification
     logger.debug(f'CSV file path: {input_file_path}')
     logger.debug(f'Excel file path: {output_file_path}')
@@ -35,17 +34,28 @@ try:
 
     logger.info(f'Excel file created at: {output_file_path}')
 
+
+def read_excel_file(excel_file_path):
+
     # Read the Excel file back into a DataFrame
-    df_excel = pd.read_excel(output_file_path)
+    df_excel = pd.read_excel(excel_file_path)
 
     # Debug: Output information about the Excel DataFrame
     logger.debug(f'Excel DataFrame shape: {df_excel.shape}')
     logger.debug(f'Excel DataFrame columns: {df_excel.columns}')
 
 
-except FileNotFoundError as e:
-    logger.error(f'File not found: {e}', exc_info=True)
-except PermissionError as e:
-    logger.error(f'Permission error: {e}', exc_info=True)
-except Exception as e:
-    logger.error(f'An unexpected error occurred: {e}', exc_info=True)
+def main():
+    # Define variables for file paths
+    input_file_path = r'D:\heart.csv'
+    output_file_path = r'D:\heart-excel.xlsx'
+
+    # Process CSV and create Excel file
+    read_csv_and_create_excel(input_file_path, output_file_path)
+
+    # Process the created Excel file
+    read_excel_file(output_file_path)
+
+
+if __name__ == "__main__":
+    main()
